Change how command responses and callbacks are handled.
The parsing of command responses has been moved from the connection
class to the client class.

Command callbacks passed to a client instance are now wrapped by
an outer callback that allows for storing the actual client instance
that is returned as the third argument. The command that generated
the response is now passed as the second argument of the callback.

This allows for a better separation of the client abstraction from
the low-level connection class and facilitates writing user code by
removing some ugly use() directives from anonymous functions.