Make dependency on phpiredis optional, default on clue/redis-protocol.
The client will use the usual phpiredis-based connection backend when
the extension is loaded, otherwise it will automatically fall back to
a different backend based on the clue/redis-protocol library.

Developers can force the client to use the pure-php connection backend
even when phpiredis is loaded simply by setting the "phpiredis" client
option to "false".